Feat:add some obj and finish camera class

+++ lib/parse-utils/objParse.ts
@@ -230,7 +230,7 @@ namespace Utils{
             var vertices = new Float32Array(numVertices * 3);
             var normals = new Float32Array(numVertices * 3);
             var colors = new Float32Array(numVertices * 4);
-            var indices = new Uint8Array(numIndices);
+            var indices = new Uint16Array(numIndices);
 
             // Set vertex, normal and color
             var index_indices = 0;
